% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterData.R
\name{FilterData}
\alias{FilterData}
\title{Title FilterData
This function removes contaminant peptides and untrusted peptides from the initialized data list.}
\usage{
FilterData(
  Peptide,
  type,
  A,
  B,
  Contamination = "CON",
  rmKRT = T,
  rmTRY = T,
  NAThres = 0.6
)
}
\arguments{
\item{Peptide}{Quantitative peptide information generated by Initiation function}

\item{type}{a vector describe the subtype of samples}

\item{A}{A type name of cells for comparison}

\item{B}{A type name of cells for comparison}

\item{Contamination}{Accession markers that annotate a contaminant peptide}

\item{rmKRT}{T or F, Whether to remove keratin protein or not}

\item{rmTRY}{T or F, Whether to remove trypsin protein or not}

\item{NAThres}{Maximum percentage of missing values allowed for each peptide}
}
\value{
A data list of peptide information without untrusted information
}
\description{
Title FilterData
This function removes contaminant peptides and untrusted peptides from the initialized data list.
}
\examples{
Peptide<-Initiation(test_raw)
Peptide<-FilterData(Peptide,type=test_type,'A','B')
}
