% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initiation.R
\name{Initiation}
\alias{Initiation}
\title{Title Initialization
This function initialize the input data to enable downstream analysis}
\usage{
Initiation(data)
}
\arguments{
\item{data}{a .Rdata file read from the peptide-level searching result
should contain columns "Accession" "Description" "Sequence" "Modifications"
Columns for peptide abundance should be annotated with "Abundance"
Better contain columns "RT" and "m/z"}
}
\value{
A data list of peptide information for downstream analysis
}
\description{
Title Initialization
This function initialize the input data to enable downstream analysis
}
\examples{
Peptide<-Initiation(test_raw)
}
