% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{Normalization}
\alias{Normalization}
\title{Title Normalization
This optinal functions performs median normalization among samples}
\usage{
Normalization(Peptide, method = "Median", group = F)
}
\arguments{
\item{Peptide}{Quantitative peptide without untrusted data and contaminant data, and without outlier samples}

\item{method}{"Median" or "Mean", indicating to normalize by median value or mean value.}

\item{group}{T or F, normalizing the peptide abundances within each group or between two sample groups.}
}
\value{
data list of normalized peptide information
}
\description{
Title Normalization
This optinal functions performs median normalization among samples
}
\examples{
Peptide<-Initiation(test_raw)
Peptide<-FilterData(Peptide,type=test_type,'A','B')
Peptide<-FilterSample(Peptide)
Peptide<-Normalization(Peptide)
}
