% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeptideRatio.R
\name{PeptideRatio}
\alias{PeptideRatio}
\title{Title PeptideRatio
This function calculates pairwise ratio of peptides between two subpopulations}
\usage{
PeptideRatio(Peptide, numerator = Peptide$type[1])
}
\arguments{
\item{Peptide}{Quantitative peptide information without untrusted data and contaminant data, and without outlier samples}

\item{numerator}{the type of cells that play the numerator when calculating the peptide ratio}
}
\value{
A data list of peptide information with a table of peptide DE-scores
}
\description{
Title PeptideRatio
This function calculates pairwise ratio of peptides between two subpopulations
}
\examples{
Peptide<-Initiation(test_raw)
Peptide<-FilterData(Peptide,type=test_type,'A','B')
Peptide<-FilterSample(Peptide)
Peptide<-Normalization(Peptide)
Peptide<-PeptideRatio(Peptide)
}
