% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreProtein.R
\name{ScoreProtein}
\alias{ScoreProtein}
\title{Title ScoreProtein
This function calculates the protein DE-scores based on peptide DE-scores and peptide abundances}
\usage{
ScoreProtein(Peptide, pvalcutoff = 0.05)
}
\arguments{
\item{Peptide}{Quantitative peptide information with a table of peptide DE-scores}

\item{pvalcutoff}{Maximum p-value for hypothesis test for differential expressed peptides.}
}
\value{
a table of protein DE-score
}
\description{
Title ScoreProtein
This function calculates the protein DE-scores based on peptide DE-scores and peptide abundances
}
\examples{
Peptide<-Initiation(test_raw)
Peptide<-FilterData(Peptide,type=test_type,'A','B')
Peptide<-FilterSample(Peptide)
Peptide<-Normalization(Peptide)
Peptide<-PeptideRatio(Peptide)
Result<-ScoreProtein(Peptide)
}
