% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_raw.R
\docType{data}
\name{test_raw}
\alias{test_raw}
\title{Title test_raw}
\format{
a data frame with 5508 peptides for 27 samples
\describe{
\item{Sequence}{sequence of peptide}
\item{Modifications}{modifications of the peptide}
\item{Accession}{Accession of the master protein of the peptide}
\item{Description}{description or gene symbol of the master protein}
\item{RT}{average retention time of the peak of peptide}
\item{m/z}{average m/z of the peak of peptide}
}
}
\usage{
test_raw
}
\description{
A dataset
}
\keyword{datasets}
