% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_type.R
\docType{data}
\name{test_type}
\alias{test_type}
\title{Title test_raw}
\format{
An object of class \code{character} of length 21.
}
\usage{
test_type
}
\description{
a vector describing the subtype of 27 samples in test_raw
}
\keyword{datasets}
